requirements.txt
Django==4.0.4
django-allauth==0.50.0
django-fractions==2.0.0
django-markdownify==0.9.1
flake8-black==0.3.3
gunicorn==20.1.0
Pillow==9.1.1
qrcode==7.3.1
djangorestframework~=3.13.1
psutil~=5.9.1
PyYAML~=6.0
